file JVMuiClassNameSetup moved to material-ui folder to short the import path for configuring the styles
